Convert TestConstants to be an interface.
