+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/TestConstants.java
@@ -1,30 +1,30 @@
 package uk.gov.hmcts.reform.finrem.caseorchestration;
 
-public class TestConstants {
-    public static final String TEST_CASE_ID = "test.case.id";
-    public static final String TEST_USER_ID = "test.user.id";
-    public static final String TEST_ORG_ID = "test.org.id";
-    public static final String TEST_CASE_ROLE = "[ROLE]";
-    public static final String TEST_DIVORCE_CASE_NUMBER = "DD12D12345";
-    public static final String TEST_SERVICE_TOKEN = "testServiceToken";
-    public static final String AUTH_TOKEN = "Bearer eyJhbGciOiJIUzI1NiJ9";
-    public static final String CASE_ID = "1234567890";
+public interface TestConstants {
+    String TEST_CASE_ID = "test.case.id";
+    String TEST_USER_ID = "test.user.id";
+    String TEST_ORG_ID = "test.org.id";
+    String TEST_CASE_ROLE = "[ROLE]";
+    String TEST_DIVORCE_CASE_NUMBER = "DD12D12345";
+    String TEST_SERVICE_TOKEN = "testServiceToken";
+    String AUTH_TOKEN = "Bearer eyJhbGciOiJIUzI1NiJ9";
+    String CASE_ID = "1234567890";
 
-    public static final String TEST_SOLICITOR_REFERENCE = "RG-123456789";
-    public static final String TEST_SOLICITOR_NAME = "Saul Goodman";
-    public static final String TEST_SOLICITOR_EMAIL = "testSolicitor@email.com";
+    String TEST_SOLICITOR_REFERENCE = "RG-123456789";
+    String TEST_SOLICITOR_NAME = "Saul Goodman";
+    String TEST_SOLICITOR_EMAIL = "testSolicitor@email.com";
 
-    public static final String TEST_RESP_SOLICITOR_REFERENCE = "RG-987654321";
-    public static final String TEST_RESP_SOLICITOR_NAME = "Test Resp Sol";
-    public static final String TEST_RESP_SOLICITOR_EMAIL = "testRespSolicitor@email.com";
+    String TEST_RESP_SOLICITOR_REFERENCE = "RG-987654321";
+    String TEST_RESP_SOLICITOR_NAME = "Test Resp Sol";
+    String TEST_RESP_SOLICITOR_EMAIL = "testRespSolicitor@email.com";
 
-    public static final String TEST_CTSC_EMAIL = "testCtsc@email.com";
-    public static final String TEST_JUDGE_EMAIL = "testJudge@email.com";
+    String TEST_CTSC_EMAIL = "testCtsc@email.com";
+    String TEST_JUDGE_EMAIL = "testJudge@email.com";
 
-    public static final String TEST_KEY = "testKey";
-    public static final String TEST_VALUE = "testValue";
-    public static final String TEST_FORM = "testForm";
-    public static final String TEST_FORM_TYPE = "testFormType";
-    public static final String TEST_BULK_UNSUPPORTED_FORM_TYPE = "unsupportedFormType";
-    public static final String TEST_URL = "https://test.com";
+    String TEST_KEY = "testKey";
+    String TEST_VALUE = "testValue";
+    String TEST_FORM = "testForm";
+    String TEST_FORM_TYPE = "testFormType";
+    String TEST_BULK_UNSUPPORTED_FORM_TYPE = "unsupportedFormType";
+    String TEST_URL = "https://test.com";
 }
